add components:check_update, update new component translations, fixed datatype of blueprint.eva and blueprint.crit, update sort_by feature and blueprint import task to ensure equip attributes sorted correctly

+++ db/migrate/20240708185555_create_blueprints.rb
@@ -34,8 +34,8 @@ class CreateBlueprints < ActiveRecord::Migration[7.0]
       t.integer :atk
       t.integer :def
       t.integer :hp
-      t.integer :eva
-      t.integer :crit
+      t.decimal :eva
+      t.decimal :crit
 
       t.timestamps
     end
